// src/enums/translationKey.enum.ts
export enum ENavLabel {
  HOME = "home",
  ABOUT = "about",
  TECHNICAL_SKILLS = "technicalSkills",
  RESUME = "resume",
  PROJECTS = "project",
  CONTACT = "contact",
}

export enum ENavAriaLabel {
  HOME = "homeLink",
  ABOUT = "aboutLink",
  TECHNICAL_SKILLS = "technicalSkillsLink",
  RESUME = "resumeLink",
  PROJECTS = "projectsLink",
  CONTACT = "contactLink",
}

export enum ECommonAriaLabel {
  TOGGLE_NAVIGATION = "toggleNavigation",
  FORK_REPOSITORY = "forkRespository",
  FORK_BUTTON = "forkButton",
}

export enum EInterestLabel {
  WEB_DEVELOPMENT = "webDevelopment",
  FRONTEND_DEVELOPMENT = "frontendDevelopment",
  BACKEND_DEVELOPMENT = "backendDevelopment",
  TOUCH_TYPING = "touchTyping",
  DOMAIN_AND_DNS = "domainAndDns",
  PROGRAMMING = "programming",
  WEB_HOSTING = "webHosting",
  PRODUCT_MANAGEMENT = "productManagement",
}